distro/app/requirements.txt
PyQt6==6.7.1
openpyxl==3.1.2
reportlab==4.0.7

# FastAPI Web API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pyodbc>=5.0.0
